requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-htmx==1.17.3
dj-database-url==2.2.0
psycopg[binary]==3.2.13
whitenoise==6.7.0
python-dateutil==2.9.0.post0
ortools==9.12.4544
gunicorn==22.0.0
